added Javadoc and organized packages
