+++ src/main/java/APP_Business_Rules/SearchUseCase/SearchDishUseCase.java
@@ -1,7 +1,6 @@
 package APP_Business_Rules.SearchUseCase;
 
 import APP_Business_Rules.DishMenu.DishDataAccess;
-import APP_Business_Rules.SearchOutputBoundary;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -11,6 +10,11 @@ import static java.lang.Double.compare;
 import static java.lang.Integer.parseInt;
 
 public class SearchDishUseCase implements SearchInputBoundary{
+    /**
+     * Class that takes care of querying dishes and filtering out the ones that do not match the search
+     * @param searchPresenter presenter that updates the UI with the restaurants that match the search
+     * @param dataAccess class that reads the dish database
+     */
     private SearchOutputBoundary searchPresenter;
     private DishDataAccess dataAccess;
 
@@ -19,6 +23,11 @@ public class SearchDishUseCase implements SearchInputBoundary{
         this.dataAccess = dataAccess;
     }
 
+    /**
+     * Function that gets all the available dishes and keeps the ones that match the search then calls the presenter
+     * to update the UI
+     * @param searchRequestModel request model that contains all the information about the search
+     */
     public void Search(SearchRequestModel searchRequestModel){
         List<List<List<String>>> data = new ArrayList<>(dataAccess.getDish("src/main/java/Frameworks_and_Drivers/Dishes.csv").values()); //Might need to change the return type of accessData() to Arraylist<Object>
         SearchResponseModel searchResponseModel = new SearchResponseModel(new ArrayList<>(), "Dish");
