+++ src/main/java/APP_Business_Rules/SearchUseCase/SearchInputBoundary.java
@@ -1,5 +1,8 @@
 package APP_Business_Rules.SearchUseCase;
 
+/**
+ * Interface implemented by search use case presenter
+ */
 public interface SearchInputBoundary {
  public void Search(SearchRequestModel searchRequestModel);
 }
