+++ src/main/java/APP_Business_Rules/SearchUseCase/SearchOutputBoundary.java
@@ -1,7 +1,9 @@
-package APP_Business_Rules;
+package APP_Business_Rules.SearchUseCase;
 
 import APP_Business_Rules.SearchUseCase.SearchResponseModel;
-
+/**
+ * Interface implemented by search use case presenter
+ */
 public interface SearchOutputBoundary {
     void update(SearchResponseModel searchResponseModel);
 }
