+++ src/main/java/APP_Business_Rules/SearchUseCase/SearchResponseModel.java
@@ -5,6 +5,11 @@ import java.util.HashMap;
 import java.util.List;
 
 public class SearchResponseModel {
+    /**
+     * Object that represents the items that match the search
+     * @param result list of items that match the search, each item is represented by a list of strings
+     * @param type represents the type of search made
+     */
     private List<List<String>> result;
     private String type;
