+++ src/main/java/APP_Business_Rules/SearchUseCase/SearchRestaurantUseCase.java
@@ -1,6 +1,5 @@
 package APP_Business_Rules.SearchUseCase;
 
-import APP_Business_Rules.SearchOutputBoundary;
 import APP_Business_Rules.RestaurantUseCase.RestaurantDataAccess;
 
 import java.util.ArrayList;
@@ -11,6 +10,11 @@ import static java.lang.Double.compare;
 import static java.lang.Double.parseDouble;
 
 public class SearchRestaurantUseCase implements SearchInputBoundary{
+    /**
+     * Class that takes care of querying restaurants and filtering out the ones that do not match the search
+     * @param searchPresenter presenter that updates the UI with the restaurants that match the search
+     * @param dataAccess class that reads the restaurant database
+     */
     private SearchOutputBoundary searchPresenter;
     private RestaurantDataAccess dataAccess;
 
@@ -19,6 +23,11 @@ public class SearchRestaurantUseCase implements SearchInputBoundary{
         this.dataAccess = dataAccess;
     }
 
+    /**
+     * Function that gets all the available restaurants and keeps the ones that match the search then calls the presenter
+     * to update the UI
+     * @param searchRequestModel request model that contains all the information about the search
+     */
     public void Search(SearchRequestModel searchRequestModel){
         List<List<String>> data = this.dataAccess.getRes(); //Might need to change the return type of accessData() to Arraylist<Object>
         SearchResponseModel searchResponseModel = new SearchResponseModel(new ArrayList<>(), "Restaurant");
