+++ src/main/java/Interface_and_Adapters/SearchScreen/SearchScreen.form
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Interface_and_Adapters.SearchScreen">
+<form xmlns="http://www.intellij.com/uidesigner/form/" version="1" bind-to-class="Interface_and_Adapters.SearchScreen.SearchScreen">
   <grid id="27dc6" binding="ResearchPanel" layout-manager="CardLayout" hgap="0" vgap="0">
     <constraints>
       <xy x="20" y="20" width="515" height="476"/>
